+++ app/Http/Controllers/ProductController.php
@@ -11,7 +11,7 @@ class ProductController extends Controller
 {
     public function create()
     {
-
+        $this->authorize('is_admin');
         $data = [];
         $listCat = Category::all();
         $data['list'] = $listCat;
@@ -21,22 +21,26 @@ class ProductController extends Controller
 
     public function store(Request $request)
     {
-        $data = $request->only('name', 'value', 'photo', 'description', 'id_category');
-
+        $this->authorize('is_admin');
+       
         $request->validate([
             'photo' => 'required|image|mimes:jpeg,png,jpg,gif,svg|max:2048',
             'name' => 'required',
-            'value' => 'required',
+            'value' => 'required', 'regex:/^\d+(\,\d{1,2})?$/',
+            'quantity' => 'required',
             'description' => 'required',
             'id_category' => 'required',
         ]);
+        $value = str_replace(['.', ','], ['', '.'], $request->input('value'));
+        $data = $request->only('name', 'value', 'photo', 'description', 'quantity', 'id_category');
+        $data['value'] = $value;
 
-        if ($request->has('photo') && $request->photo->isValid()) { 
-        $nameImage = $request->file('photo')->getClientOriginalName();
-
-        Storage::put("/public/images/produtos/{$nameImage}", file_get_contents($request->photo));
-
-        $data['photo'] = "/storage/images/produtos/{$nameImage}";
+        if ($request->hasFile('photo')) {
+            $nameImage = time() . '_' . $request->file('photo')->getClientOriginalName();
+        
+            $path = $request->file('photo')->storeAs('images/products', $nameImage, 'public');
+    
+            $data['photo'] = "/storage/$path";
         }
         Product::create($data);
 
@@ -47,49 +51,49 @@ class ProductController extends Controller
     {
         $show = Product::find($product);
 
-        // dd($show);
         return view('products.show', compact('show'));
     }
 
     public function edit($product)
     {
-        // $this->authorize('is_admin');
+        $this->authorize('is_admin');
+        $catList = Category::all();
+        $prod = Product::find($product);
 
-        // lista Categorias
-        $catList = [];
-        $cat = Category::all();
-        $catList['cat'] = $cat;
-
-        //lista Produtos
-        $edit = Product::find($product);
-
-        return view('products.edit', compact('edit'), $catList);
+        return view('products.edit', compact('prod', 'catList'));
     }
 
-    public function update($product, Request $request)
-    {
-        // $this->authorize('is_admin');
-
-        $data = $request->all();
-
-        //upload IMAGE
-        if ($request->has('photo') && $request->photo->isValid()) {
-
-            $nameImage = $request->file('photo')->getClientOriginalName();
-
-            Storage::put("/public/images/produtos/{$nameImage}", file_get_contents($request->photo));
-
-            $data['photo'] = "/storage/images/produtos/{$nameImage}";
-        }
-
-        Product::find($product)->update($data);
-
-        return redirect()->route('users.index');
+    public function update(Request $request, $product)
+{
+    $this->authorize('is_admin');
+    
+    $request->validate([
+        'photo' => 'sometimes|image|mimes:jpeg,png,jpg,gif,svg', 
+        'name' => 'required|string|max:255',
+        'value' => ['required', 'regex:/^\d+(\,\d{1,2})?$/'], 
+        'quantity' => 'required|integer|min:1',
+        'description' => 'required|string|max:1000',
+        'id_category' => 'required|integer|exists:categories,id',
+    ]);
+    
+    $value = str_replace(['.', ','], ['', '.'], $request->input('value'));
+    $data = $request->all();
+    $data['value'] = $value;
+    if ($request->hasFile('photo')) {
+        $nameImage = time() . '_' . $request->file('photo')->getClientOriginalName();
+        $path = $request->file('photo')->storeAs('images/products', $nameImage, 'public');
+        $data['photo'] = "/storage/$path";
     }
+    
+    Product::findOrFail($product)->update($data);
+
+    return redirect()->route('users.index')->with('success', 'Produto atualizado com sucesso!');
+}
+
 
     public function destroy($product)
     {
-        // $this->authorize('is_admin');
+        $this->authorize('is_admin');
 
         Product::findOrFail($product)->delete();
